draft: add countingsort.c3
NOTE: BUGGED! DON'T MERGE AS IS

There's bizarre read access errors for `insertionsort_indexed` and `quicksort_indexed` specifically inside `csort`. `csort` on its own does not appear to have a problem, and in my testing `insertionsort_indexed` and `quicksort_indexed` outside works perfectly fine EVEN if given out of bounds indexs. So I'm at a total loss as to what is wrong. My best guess is it's something to do with `list`'s lifetime, but I don't understand how macros work so, I'll leave this here as an example. If you want to see it "work" (but be slow) remove the fallbacks to quicksort and insertionsort and just let `csort` handle all cases.