sort: make countingsort.c3 even more branchless
Track even more specific partitions, such that we can define loops without if-statements within them.

If there's a branchless-ish way to do a 3 way partition of the provided indexs that'd be perfect. Here we're using a partition algorithm to split which indexs use which fallback.

May be better to iterate over the counts a second time?