fix quicksort stalling
Fix quicksort stalling by passing `#list` instead of `list` to the
`@partition` function. The stalling occurs when a fixed-sized array is
passed to quicksort. Fixed-sized arrays parameters will be passed by
value (and not by references); hence, a fixed-sized array will not be
sorted but it should also not stall the algorithm.

As additional measure (or alternative), we could use a contract to avoid
passing fixed-size arrays or vectors to the quicksort function, i.e.
```
@require values::@typekind(list) != ARRAY
@require values::@typekind(list) != VECTOR
```

Fixes: #1845